% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{cal.SINBA.Activity}
\alias{cal.SINBA.Activity}
\title{cal.SINBA.Activity}
\usage{
cal.SINBA.Activity(
  SINBA_originalID = NULL,
  target_list = NULL,
  cal_mat = NULL,
  es.method = "weightedmean",
  std = TRUE
)
}
\arguments{
\item{SINBA_originalID}{character vector or the return data frame generated by \code{combineNet2target_list} with return_type="SINBA_originalID". \code{SINBA_originalID} contains the seed-partner driver pairs; seed gene name and partner gene name are seperated by colon. \code{SINBA_originalID} is optional.}

\item{target_list}{list, the driver-to-target list object.\code{target_list} is generated by \code{get.combined.network} and contains all the driver gene regulons.}

\item{cal_mat}{numeric matrix, the expression matrix of genes/transcripts.}

\item{es.method}{character, method applied to calculate the activity value. User can choose from c ("mean", "weightedmean", "maxmean", "absmean", "max" and "min"). Default is "weightedmean".}

\item{std}{logical, if TRUE, the expression matrix will be normalized by column. Default is TRUE.}
}
\description{
This function will Calculate Activity Value for Each Driver or seed-partner pairs.
}
\examples{
sinba_originalID<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN","STAT3","PTCRA"),return_type="SINBA_originalID")
cal_mat<-Biobase::exprs(SINBA.par$cal.eset)
sinba_ac_mat<-cal.SINBA.Activity(SINBA_originalID=sinba_originalID,target_list=SINBA.par$combined.network$target_list,cal_mat=cal_mat)
single_ac_mat<-cal.SINBA.Activity(target_list=SINBA.par$combined.network$target_list,cal_mat=cal_mat)
\dontrun{
}
}
