% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{ChEMBL_drugphase_levels}
\alias{ChEMBL_drugphase_levels}
\title{ChEMBL_drugphase_levels}
\usage{
ChEMBL_drugphase_levels()
}
\description{
the drug phases used in ChEMBL. Users can select sepecific drug phase
}
\examples{
ChEMBL_drugphase_levels()
}
