% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{combine_names}
\alias{combine_names}
\title{combine_names}
\usage{
combine_names(x, target_list, partner_name)
}
\arguments{
\item{x}{seed gene name which is included in target list}

\item{target_list}{\code{target_list} is a list contain all the regulons}

\item{partner_name}{partner genes for combination}
}
\value{
Return a comination name list. seed and partner genes are seperated by colon.
}
\description{
combine_names
}
