% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{combineNet2target_list}
\alias{combineNet2target_list}
\title{combineNet2target_list}
\usage{
combineNet2target_list(
  target_list = NULL,
  seed_name = NULL,
  partner_name = NULL,
  return_type = "SINBA_originalID"
)
}
\arguments{
\item{target_list}{list, the driver-to-target list object.\code{target_list} is generated by \code{get.combined.network} and contains all the driver gene regulons.}

\item{seed_name}{the driver genes will be used as seed for combination, must be included the the target_list}

\item{partner_name}{optional, the partner driver genes will be used as partner for combination. The default setting will use all the other drivers except the seed gene in the target_list.}

\item{return_type}{character,the type of return data. User can choose from "SINBA_target_list" and "SINBA_originalID". Default is "SINBA_originalID", which returns the combination originalID name for downstream activity calculation, "SINBA_target_list" returns the full combination target list.}
}
\description{
This function will create a combined target list of seed-partner driver pairs.
}
\examples{
out_net_file1<-"path1_to_network_file/consensus_network_ncol_.txt"
out_net_file2<-"path2_to_network_file/consensus_network_ncol_.txt"
net1  <- get.single.network(network_file=out_net_file1)
net2  <- get.single.network(network_file=out_net_file2)
combined.network<-get.combined.network(network1=net1,network2=net2)
sinba_target_list<-combineNet2target_list(target_list=combined.network$target_list,seed_name="LCK")
sinba_target_list<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN"),return_type="SINBA_target_list")
sinba_originalID<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN"),return_type="SINBA_originalID")
\dontrun{
}
}
