% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{draw.box_SINBA}
\alias{draw.box_SINBA}
\title{draw.box_SINBA}
\usage{
draw.box_SINBA(
  ac_mat_sinba = NULL,
  ac_mat_single = NULL,
  SINBA_originalID = NULL,
  phenotype_info = NULL,
  group_col = NULL,
  group_levels = NULL,
  group_colors = NULL,
  main = ""
)
}
\arguments{
\item{ac_mat_sinba, }{numeric matrix, the activity values of the selected seed-partner driver across all samples.}

\item{ac_mat_single, }{numeric matrix, the activity values of single driver across all samples}

\item{SINBA_originalID, }{seed and parterne driver combination, is generated by \code{combineNet2target_list} with return_type="SINBA_originalID"}

\item{phenotype_info, }{a data frame with sample group information}

\item{group_col, }{character, the column name to be used for group}

\item{group_levels, }{a vector of characters, the order of the group}

\item{group_colors, }{a vector of characters, colors to be used for each group}

\item{main, }{character, the main title of the plot}
}
\description{
draw.box_SINBA draws a scatter box plot to visualize one selected seed partner driver single and combined activity values across different phenotype subgroups of samples. Three side-by-side scatter box plots will be created. The left plot shows combined driver paire activity values in different phenotype subgroups, each point is a sample. The middle and right plot shows seed and partner driver activity values in different phenotype subgroups, each point is a sample.
}
\examples{
ac_mat_sinba<-exprs(SINBA.par$SINBA.ac.eset)
ac_mat_single<-exprs(SINBA.par$SINBA_single.ac.eset)
SINBA_originalID<-"LCK:STAT3"
phenotype_info<-pData(SINBA.par$SINBA.ac.eset)
phenotype_info$condition<-ifelse(phenotype_info$das_lc50>80,"Resistant","Sensitive")
g<-draw.box_SINBA(ac_mat_sinba=ac_mat_sinba,ac_mat_single=ac_mat_single,SINBA_originalID=SINBA_originalID,phenotype_info=phenotype_info,group_col="condition",group_levels=c("Sensitive","Resistant"),group_colors=c("red","blue"))
}
