% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{draw.GSEA_SINBA}
\alias{draw.GSEA_SINBA}
\title{draw.GSEA_SINBA}
\usage{
draw.GSEA_SINBA(
  rank_profile = NULL,
  use_genes_SINBA = NULL,
  use_genes_seed = NULL,
  use_genes_partner = NULL,
  use_direction = TRUE,
  use_direction_SINBA = NULL,
  use_direction_seed = NULL,
  use_direction_partner = NULL,
  main = "",
  pdf_file = NULL,
  annotation = NULL,
  annotation_cex = 1.2,
  left_annotation = NULL,
  right_annotation = NULL
)
}
\arguments{
\item{rank_profile}{a named vector of numerics, the differential values (DE or DA) calculated from a sample comparison. Names of the vector must be gene names.The differential values can be "logFC", "t-statistics" or "Z-statistics".}

\item{use_genes_SINBA}{a vector of characters, a vector of genes to display. The genes the targe genes from seed-partner pair. Seed-partner regulon can be generated by \code{combineNet2target_list} with (return_type="SINBA_target_list")}

\item{use_genes_seed}{a vector of characters, a vector of genes to display. The genes the targe genes from the seed driver.}

\item{use_genes_partner}{a vector of characters, a vector of genes to display. The genes the targe genes from the partner driver.}

\item{use_direction}{a logical value, if TRUE, The target direction values need provided by user. The default value is TRUE.}

\item{use_direction_SINBA}{a vector of numeric 1s and -1s, 1 is positive regulation from the seed-partner pair, -1 is negative regulation from seed-partner pair. Users can get this vector by converting the signs of "spearman". If NULL, no regulation direction will be displayed. Default is NULL.}

\item{use_direction_seed}{a vector of numeric 1s and -1s, 1 is positive regulation from the seed, -1 is negative regulation from seed. Users can get this vector by converting the signs of "spearman". If NULL, no regulation direction will be displayed. Default is NULL.}

\item{use_direction_partner}{a vector of numeric 1s and -1s, 1 is positive regulation from the partner, -1 is negative regulation from partner. Users can get this vector by converting the signs of "spearman". If NULL, no regulation direction will be displayed. Default is NULL.}

\item{main}{character, an overall title for the plot. Default is "".}

\item{pdf_file}{character, the file path to save as PDF file. If NULL, no PDF file will be saved. Default is NULL.}

\item{annotation}{character, the annotation set by users for easier reference.
Normally the annotation is the P-value or other statistics to show the significance of the interested seed-partner driver pairs.If NULL, will perform a Kolmogorov-Smirnov test to get the significance value. Default is NULL.}

\item{annotation_cex}{numeric, giving the amount by which the text of annotation should be magnified relative to the default. Default is 1.2.}

\item{left_annotation}{character, annotation displayed on the left of the figure, representing left condition of the \code{rank_profile}. Default is "".}

\item{right_annotation}{character, annotation displayed on the right of the figure, representing right condition of the \code{rank_profile}. Default is "".}
}
\value{
Return a logical value. If TRUE, the plot has been created successfully.
}
\description{
GSEA (Gene Set Enrichment Analysis) Plot for seed, partner and combined Seed-partner regulons.
}
\examples{
lck_fyn<-combineNet2target_list(target_list=SINBA.par$sinba.network$target_list,seed_name="LCK",partner_name="FYN",return_type="SINBA_target_list")
genes_SINBA<-lck_fyn[[1]]$target
genes_seed<-SINBA.par$sinba.network$target_list$LCK$target
genes_partner<-SINBA.par$sinba.network$target_list$FYN$target
dir_SINBA<-sign(lck_fyn[[1]]$spearman)
dir_seed<-sign(SINBA.par$sinba.network$target_list$LCK$spearman)
dir_partner<-sign(SINBA.par$sinba.network$target_list$FYN$spearman)
rank_profile <- SINBA.par$DE[[comp_name]]$`Z-statistics`
names(rank_profile)<-SINBA.par$DE[[comp_name]]$ID
rank_profile<-rank_profile[order(-rank_profile)]
draw.GSEA_SINBA(rank_profile=rank_profile,use_genes_SINBA=genes_SINBA,use_genes_seed=genes_seed,use_genes_partner=genes_partner,use_direction=TRUE, use_direction_SINBA=dir_SINBA,use_direction_seed=dir_seed,use_direction_partner=dir_partner,main="LCK_FYN",pdf_file=sprintf("\%s/GSEA_LCK_FYN.pdf",plot.dir),annotation=NULL,annotation_cex=1.2,left_annotation="Sensitive",right_annotation="Resistant")

\dontrun{
}

}
