% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{Drugbank_drugphase_levels}
\alias{Drugbank_drugphase_levels}
\title{Drugbank_drugphase_levels}
\usage{
Drugbank_drugphase_levels()
}
\description{
the drug phases used in drugbank. Users can select sepecific drug phase
}
\examples{
Drugbank_drugphase_levels()
}
