% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{drugdatabase_names}
\alias{drugdatabase_names}
\title{drugdatabase_names}
\usage{
drugdatabase_names()
}
\description{
the curated drug databases. Users can select sepecific drug database
}
\examples{
drugdatabase_names()
}
