% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{drugdb.preload}
\alias{drugdb.preload}
\title{Preload drug database files into R workspace for SINBA}
\usage{
drugdb.preload(db = NULL, main.dir = NULL, out.dir = NULL)
}
\arguments{
\item{db}{A vector of characters, users can choose from "BPdb","ChEMBL","cMAP","Drugbank_V5.1", "rDGIdb_1.8.0" and "Repurposing_HUB". If NULL, all the drug databases will be used. Default is NULL.}

\item{main.dir}{character, the main directory for SINBA package.
If NULL, will be \code{system.file(package = "SINBA")}. Default is NULL.}

\item{out.dir}{charater, the directory to save drug database information. If NULL, will be \code{system.file(package = "SINBA")}. Default is NULL.}
}
\description{
\code{drugdb.preload} is a pre-processing function for SINBA. It preloads needed drug databases into R workspace,and saves it locally under data/ directory.
}
\details{
Users need to choose the drug databse sources(e.g. "BPdb","ChEMBL","cMAP","Drugbank_V5.1", "rDGIdb_1.8.0","Repurposing_HUB")
}
\examples{
drugdb.preload()

\dontrun{
drugdb.preload()
}
}
