% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_fixed_matrix}
\alias{fit_fixed_matrix}
\title{Maximum Likelihood Estimation of Births With Fixed Transition Matrix}
\usage{
fit_fixed_matrix(
  tree,
  data,
  rate_mat,
  model = NULL,
  pi_x = NULL,
  pi_y = NULL,
  root = NULL,
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{pi_x}{The transition probability at the birth of x trait.
If NULL it will set 1.0 for the state 1.}

\item{pi_y}{The transition probability at the birth of y trait.
If NULL it will set 1.0 for the state 1.}

\item{root}{Root prior probabilities.
By default,
it uses a FitzJohn prior.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_fixed_matrix()} searches for the maximum likelihood estimate
of birth events
given a transition matrix
under the Sinba model.
}
