% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_mixed}
\alias{fit_mixed}
\title{Maximum Likelihood Estimation With Mixed Births}
\usage{
fit_mixed(
  tree,
  data,
  model = NULL,
  trait = "x",
  pi_trait = NULL,
  root = NULL,
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{trait}{The free trait
(i.e., the birth to be estimated).
By default is "x",
for the first trait.
Use "y" for the second trait.}

\item{pi_trait}{The transition probability for the birth
of the free trait.}

\item{root}{Root prior probabilities.
By default it use FitzJohn et al. (2009) method,
in which ancestral states are weighted by its own likelihood.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_mixed()} searches for the maximum likelihood estimate
with the Sinba model assuming that one of the births
is fixed at the root,
and the other is free.
It estimates the values of the parameters for the Q matrix
as well as the location of the birth event
of the free trait.
}
