% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{fit_pagel}
\alias{fit_pagel}
\title{Maximum Likelihood Estimation of the Pagel's Model Under the Sinba Model}
\usage{
fit_pagel(tree, data, model = NULL, opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_pagel()} searches for the maximum likelihood estimate
of a given model for two traits
using the Pagel (1994) model.
}
