% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_simultaneous}
\alias{fit_simultaneous}
\title{Maximum Likelihood Estimation With a Simultaneous Birth}
\usage{
fit_simultaneous(
  tree,
  data,
  model = NULL,
  pi_xy = NULL,
  root = NULL,
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{pi_xy}{The transition probability at the birth
of x and y traits.
If NULL it will set 1.0 for the state 11.}

\item{root}{Root prior probabilities.
By default is uses FitzJohn prior.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_simultaneous()} searches for the maximum likelihood estimate
with the Sinba model assuming that the birth of two traits
is simultaneous.
It estimate the values of the parameters for the Q matrix
as well as the birth event.
}
