% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_sinba}
\alias{fit_sinba}
\title{Maximum Likelihood Estimation of the Sinba Model}
\usage{
fit_sinba(tree, data, root = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
A second,
and optionally a third column,
should contain the data,
coded as 0 and 1.}

\item{root}{Root prior probabilities.}
}
\description{
\code{fit_sinba()} searches for the maximum likelihood estimate
using the Sinba model
for one or two traits.
}
