% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.r
\name{fit_sinba_single}
\alias{fit_sinba_single}
\title{Maximum Likelihood Estimation of the Sinba Model in a Single Trait}
\usage{
fit_sinba_single(tree, data, model = NULL, root = NULL, opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
A second should contain the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()}
or \code{new_hidden_model()}.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_sinba_single()} searches for the maximum likelihood estimate
with the Sinba model
for a single trait.
}
