% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{fitness_function}
\alias{fitness_function}
\title{fitness_function}
\usage{
fitness_function(solution, matrix, n, m)
}
\arguments{
\item{solution}{a vector of n+m numbers.}

\item{matrix}{the original matrix to optimize.}

\item{n}{row numbers.}

\item{m}{column numbers.}
}
\description{
Fitness function to calculate the SUM of the selected submatrix
}
