% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fixed_sinba}
\alias{fixed_sinba}
\title{Estimate the Likelihood for a Fixed Transition Matrix
and Fixed Births}
\usage{
fixed_sinba(
  tree,
  data,
  rate_mat,
  births,
  model = NULL,
  pi_x = NULL,
  pi_y = NULL,
  root = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{births}{A list with one or two element,
each element being a list that define a birth event,
with a fields \code{node} indicating the birth of the trait
and \code{age} indicating the time from the start of the edge
in which the event happens.
If the model is for two traits,
and there is single defined birth,
the same birth will be used for both traits.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{pi_x}{The transition probability at the birth of x trait.
If NULL it will set 1.0 for the state 1.}

\item{pi_y}{The transition probability at the birth of y trait.
If NULL it will set 1.0 for the state 1.}

\item{root}{Root prior probabilities.
By default it use FitzJohn et al. (2009) method,
in which ancestral states are weighted by its own likelihood.}
}
\description{
\code{fixed_sinba()} calculates the likelihood
for a fixed transition matrix
and fixed birth events
under the Sinba model.
}
