% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.r
\name{fixed_sinba_single}
\alias{fixed_sinba_single}
\title{Estimate the Likelihood for a Given Sinba Transition Matrix in a Single Trait}
\usage{
fixed_sinba_single(tree, data, rate_mat, birth, model = NULL, root = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
A second should contain the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the trait.}

\item{birth}{A list with the field \code{node}
indicated the birth of the trait,
and \code{age} the time from the start of the edge
in which the event happens.}

\item{model}{A model build with \code{new_model()}
or \code{new_hidden_model()}.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}
}
\description{
\code{fixed_sinba_single()} calculates the likelihood
for a given transition matrix
for a single trait
under the Sinba model.
}
