% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{generate.AUCTable}
\alias{generate.AUCTable}
\title{generate.AUCTable}
\usage{
generate.AUCTable(pair.AUC, single.AUC, use_comp)
}
\arguments{
\item{pair.AUC}{a list of results from \code{ROCTest} with seed-partner combined activities.}

\item{single.AUC}{a list of results from \code{ROCTest} with singe driver activities.}

\item{use_comp}{a vector of characters, the name of multiple comparisons. It will be used to name the columns of AUC table.}
}
\description{
The master table gathers all the results of \code{ROCTest}from multiple comparisons.
}
\examples{
pair.AUC<-list()
single.AUC<-list()
for(i in 1:dim(comps)[1]){
 comp_name <- sprintf('\%s.Vs.\%s',comps[i,1],comps[i,2]) ## each comparison must give a name !!!
 G0  <- rownames(phe_info)[which(phe_info$group==comps[i,2])] # get sample list for G0
 G1  <- rownames(phe_info)[which(phe_info$group==comps[i,1])] # get sample list for G1
 pair.AUC[[comp_name]]<-ROCTest(eset=SINBA.par$SINBA_AC.eset,G0=G0,G1=G1)
 single.AUC[[comp_name]]<-ROCTest(eset=SINBA.par$single_AC.eset,G0=G0,G1=G1)}
 use_comp<-names(pair.AUC)
auc_tab<-generate.AUCTable(pair.AUC = pair.AUC,single.AUC = single.AUC,use_comp = use_comp)
\dontrun{
}
}
