% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{get.combined.network}
\alias{get.combined.network}
\title{get.combined.network}
\usage{
get.combined.network(network1 = NULL, network2 = NULL)
}
\arguments{
\item{network1}{network object, which is generated by \code{get.single.network} contaning \code{network_data}. \code{network_dat} is a data.frame, contains all the information of the network SJARACNe constructed.\code{network_dat} is a data.frame, contains all the information of the network SJARACNe constructed and must contain four columns with column names "source" (driver) and "target" (target genes), "MI" (mutual information) and "spearman" (spearman correlation coefficient).}

\item{network2}{network object, which is generated by \code{get.single.network} contaning \code{network_data}. \code{network_dat} is a data.frame, contains all the information of the network SJARACNe constructed and must contain four columns with column names "source" (driver) and "target" (target genes), "MI" (mutual information) and "spearman" (spearman correlation coefficient).}
}
\description{
This function will merge two networks into one network prepared for SINBA activity calculation.
}
\examples{
out_net_file1<-"path1_to_network_file/consensus_network_ncol_.txt"
out_net_file2<-"path2_to_network_file/consensus_network_ncol_.txt"
net1  <- get.single.network(network_file=out_net_file1)
net2  <- get.single.network(network_file=out_net_file2)
combined.network<-get.combined.network(network1=net1,network2=net2)
\dontrun{
}
}
