% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{interactionTypes}
\alias{interactionTypes}
\title{interactionTypes}
\usage{
interactionTypes()
}
\description{
the available interactionTypes used in DGIdb. Users can select specific interactionTypes
}
\examples{
interactionTypes()
}
