% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-sinba.r
\name{is_full_sinba}
\alias{is_full_sinba}
\title{Return true if the data set is a full Sinba dataset}
\usage{
is_full_sinba(tree, data, model = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{A}{data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}
}
\description{
\code{is_full_sinba()} checks if a dataset is a full Sinba dataset,
that is,
if for each trait at least one state born
at a node different from the root,
and,
if there are two traits,
they form a nested set.
}
