% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{is_independent}
\alias{is_independent}
\title{Test If a Model is an Independent Model}
\usage{
is_independent(model)
}
\arguments{
\item{model}{A model definition as built by ´new_model()´.}
}
\description{
\code{is_independent()} returns true if a simple two trait model
(i.e., without hidden states)
using the Pagel conditions.
}
