% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{l2l_ftest}
\alias{l2l_ftest}
\title{l2l_ftest}
\usage{
l2l_ftest(gsc1, gsc2, N)
}
\arguments{
\item{gsc1}{a list of genesets, each geneset contains multiple gene names.}

\item{gsc2}{a list of genesets, each geneset contains multiple gene names.}

\item{N}{mumeric, total unique gene number in gsc1 and gsc2.}
}
\value{
Return a table of all fisher exact testing results.
}
\description{
l2l_ftest
}
