% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{logLik.fit_mk}
\alias{logLik.fit_mk}
\title{Extract Log-Likelihood from a "fit_mk" Object}
\usage{
\method{logLik}{fit_mk}(object, ...)
}
\arguments{
\item{object}{An object of type "fit_mk".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{logLik} method
on a "fit_mk" object.
}
