% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{logLik.fit_pagel}
\alias{logLik.fit_pagel}
\title{Extract Log-Likelihood From a "fit_pagel" Object}
\usage{
\method{logLik}{fit_pagel}(object, ...)
}
\arguments{
\item{object}{An object of type "fit_pagel".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{logLik} method
on a "fit_pagel" object.
}
