% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_fit.r
\name{logLik.fit_sinba}
\alias{logLik.fit_sinba}
\title{Extract Log-Likelihood From a "fit_sinba" Object}
\usage{
\method{logLik}{fit_sinba}(object, ...)
}
\arguments{
\item{object}{An object of type "fit_sinba".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{logLik} method
on a "fit_sinba" object.
}
