% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.r
\name{logLik.fit_sinba_births}
\alias{logLik.fit_sinba_births}
\title{Extract Log-Likelihood From a "fit_sinba_births" Object}
\usage{
\method{logLik}{fit_sinba_births}(object, ...)
}
\arguments{
\item{object}{An object of type "fit_sinba_births".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{logLik} method
on a "fit_sinba_births" object.
}
