% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.r
\name{logLik.fit_sinba_single}
\alias{logLik.fit_sinba_single}
\title{Extract Log-Likelihood From a "fit_sinba_single" Object}
\usage{
\method{logLik}{fit_sinba_single}(object, ...)
}
\arguments{
\item{object}{An object of type "fit_sinba_single".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{logLik} method
on a "fit_sinba_single" object.
}
