% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map_pagel}
\alias{map_pagel}
\title{Stochastic Map For the Pagel´s Model}
\usage{
map_pagel(fitted, n = 100)
}
\arguments{
\item{fitted}{An object of the type 'fit_mk'
(i.e., a reconstruction produced with \code{fit_pagel()} function).}

\item{n}{Number of stochastic maps.
Default is 100.}
}
\description{
\code{map_pagel()} build one or more stochastic mappings
using a reconstruction from the Pagel's model.
}
