% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map_sinba}
\alias{map_sinba}
\title{Stochastic Map For the Sinba Model}
\usage{
map_sinba(fitted, n = 100)
}
\arguments{
\item{fitted}{An object of the type 'fit_sinba'
(i.e., a reconstruction using the Sinba model).}

\item{n}{Number of stochastic maps.
Default is 100.}
}
\description{
\code{map_sinba()} uses a reconstruction using the Sinba model
to build one or more stochastic mappings.
}
