% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_add}
\alias{model_add}
\title{Add a New Parameter to a Model}
\usage{
model_add(model, cell)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}

\item{cell}{A vector with the row and column of the new parameter.}
}
\description{
\code{model_add()} reads a model definition
and add a new parameter at the indicated cell
of the model matrix.
}
