% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_as}
\alias{model_as}
\title{Transforms a Model Using a Defined Model}
\usage{
model_as(model, def)
}
\arguments{
\item{model}{The model to be updated.}

\item{def}{The model definition as used in \code{new_model()}.}
}
\description{
\code{model_as()} reads a model and transforms it
into one of the model defined by \code{new_model()}.
It only works on models without hidden rates.
The function equates all the parameters of the model
with the parameters of the of the definition,
and set all transformations between hidden states
as a single parameter.
}
