% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_collapse}
\alias{model_collapse}
\title{Removes Parameters for Unobserved States.}
\usage{
model_collapse(model, tree, data)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
The model must be defined for two traits.}

\item{tree}{A phylogenetic trees of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}
}
\description{
\code{model_collapse()} removes all unobserved states
from a model matrix.
It returns an error if the removal produces
a full inactive matrix.
}
