% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_drop}
\alias{model_drop}
\title{Remove a Parameter From a Model}
\usage{
model_drop(model, params)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}

\item{params}{A vector with the parameter IDs to be removed.}
}
\description{
\code{model_drop()} removes one o more parameters
from a model.
}
