% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_equate}
\alias{model_equate}
\title{Make Equivalent Two or More Parameters}
\usage{
model_equate(model, params)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}

\item{params}{A vector with the parameter IDs to be equated.}
}
\description{
\code{model_equate()} read a model definition
and make the indicated parameters
to be identical.
}
