% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_equate_trans_rates}
\alias{model_equate_trans_rates}
\title{Make Equivalent all Transition Rates}
\usage{
model_equate_trans_rates(model)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}
}
\description{
\code{model_equate_trans_rates()} read a mode definition
and make all transition rates
(i.e., a model created with \code{new_rates_model()})
equal.
}
