% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_join}
\alias{model_join}
\title{Join Two Models into a Single Model}
\usage{
model_join(x, y)
}
\arguments{
\item{x}{The model that represents the first trait.}

\item{y}{The model that represents the second trait.}
}
\description{
\code{model_join()} takes to models of a single trait
and merged it as a single Pagel-like model.
}
