% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_set}
\alias{model_set}
\title{Set a Parameter to a Particular Model Cell}
\usage{
model_set(model, cell, id)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}

\item{cell}{A vector with the row and column of the new parameter.}

\item{id}{The ID of the parameter.}
}
\description{
`model_set()`` sets the parameter ID for a particular cell
of the model matrix.
If the ID is zero,
it will remove the parameter ID of the indicated cell.
}
