% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{model_to_ard}
\alias{model_to_ard}
\title{Transforms a Model into an ARD model}
\usage{
model_to_ard(model)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}
}
\description{
\code{model_to_ard()} reads a model
and transforms it into an ARD model.
}
