% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{new_hidden_model}
\alias{new_hidden_model}
\title{Create a New Model With Hidden States}
\usage{
new_hidden_model(states, traits = 2)
}
\arguments{
\item{states}{A list with one or more elements,
each element containing a trait state
and its hidden states.}

\item{traits}{The number of traits in the model.
By default it is two.
It can also be a single trait.}
}
\value{
A model matrix with hidden states.
}
\description{
\code{new_hidden_model()} creates a new model matrix
with hidden states.
The hidden states are defined in a list
with three elements:
\code{trait} to indicates the trait
(either "x", the first trait,
or "y", the second trait);
\code{state} to indicate the observed state
(either 0 or 1);
and \code{hidden} with a vector of the labels
for the hidden states
(e.g. "a", "b", "c").
By default,
the model is the maximal independent model
for the given states.
}
