% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{new_model}
\alias{new_model}
\title{Build a Model Matrix}
\usage{
new_model(model = "")
}
\arguments{
\item{model}{Model of evolution for the traits.
By default it creates a matrix in which all parameters
are different,
this is the standard correlated model ("CORR").
In the independent model ("IND")
both traits evolve independently
(i.e., this is equivalent to using two Q matrices).
In the "ER" model both traits have equal rates
in any direction;
the "ER2" model also has equal rates,
but rates are different for each trait;
in the "ERs" model the rates of state transitions are equal
but can be different depending on the state.
If the "SYM" model changes between states are equal.
There a two full dependant models,
"x" for a model in which trait x
(the first trait)
depends on y;
and "y" in which trait y
(the second trait)
depends on x.
In the "sCORR" model,
rates are correlated by the state of the other trait.
In the "CMK" model,
is like the symmetrical model,
but changes of more than one state are allowed.}
}
\description{
\code{new_model()} creates a new model matrix.
}
