% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{new_rates_model}
\alias{new_rates_model}
\title{Create a New Model With Hidden Rates}
\usage{
new_rates_model(model = "", rates = NULL, traits = 2)
}
\arguments{
\item{model}{Model of evolution for the traits.
By default it returns the Pagel's independent model ("IND")
(i.e., this is equivalent to using two Q matrices).
The "CORR" model is the Pagel's correlated model,
in which each transition has a different parameter.
In the "ER" model both traits have equal rates
in any direction;
the "ER2" model also has equal rates,
but rates are different for each trait;
in the "ERs" model the rates of state transitions are equal
but can be different depending on the state.
If the "SYM" model changes between states are equal.
There a two full dependant models,
"x" for a model in which trait x
(the first trait)
depends on y;
and "y" in which trait y
(the second trait)
depends on x.
In the "sCORR" model,
rates are correlated by the state of the other trait.
In the "CMK" model,
is like the symmetrical model,
but changes of more than one state are allowed.}

\item{rates}{The labels to identify the rates.}

\item{traits}{The number of traits in the model.
By default it is two.
It can also be a single trait.}
}
\description{
\code{new_rates_model()} creates a new model matrix
with hidden rates.
}
