% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{print.fit_mk}
\alias{print.fit_mk}
\title{Basic Print For a "fit_mk" Object}
\usage{
\method{print}{fit_mk}(x, digits = 6, ...)
}
\arguments{
\item{x}{An of type "fit_mk".}

\item{digits}{The number of digits for decimal output.}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{print} method
on a \code{fit_mk} object.
}
