% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_fit.r
\name{print.fit_sinba}
\alias{print.fit_sinba}
\title{Basic Print For a "fit_sinba" Object}
\usage{
\method{print}{fit_sinba}(x, digits = 6, ...)
}
\arguments{
\item{x}{An of type "fit_sinba".}

\item{digits}{The number of digits for decimal output.}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{print} method
on a \code{fit_sinba} object.
}
