% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.r
\name{print.fit_sinba_births}
\alias{print.fit_sinba_births}
\title{Basic Print For a "fit_sinba_births" Object}
\usage{
\method{print}{fit_sinba_births}(x, digits = 6, ...)
}
\arguments{
\item{x}{An of type "fit_sinba_births".}

\item{digits}{The number of digits for decimal output.}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{print} method
on a \code{fit_sinba_births} object.
}
