% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{print.fixed_sinba}
\alias{print.fixed_sinba}
\title{Basic Print For a "fixed_sinba" Object}
\usage{
\method{print}{fixed_sinba}(x, digits = 6, ...)
}
\arguments{
\item{x}{An object of type "fixed_sinba".}

\item{digits}{The number of digits for decimal output.}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{print} method
on a \code{fixed_sinba} object.
}
