% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.r
\name{print.fixed_sinba_single}
\alias{print.fixed_sinba_single}
\title{Basic Print For a "fixed_sinba_single" Object}
\usage{
\method{print}{fixed_sinba_single}(x, digits = 6, ...)
}
\arguments{
\item{x}{An object of type "fixed_sinba_single".}

\item{digits}{The number of digits for decimal output.}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements yje \code{print} method
on a \code{fixed_sinba_single} object.
}
