% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{print.sinba_model}
\alias{print.sinba_model}
\title{Basic Print For a "sinba_model" Object}
\usage{
\method{print}{sinba_model}(x, ...)
}
\arguments{
\item{x}{An object of type "sinba_model".}

\item{...}{Additional arguments are unused.}
}
\description{
This method implements the \code{print} method
on a \code{sinba_model} object.
}
