% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{queryDgidb_V5_drug}
\alias{queryDgidb_V5_drug}
\title{queryDgidb_V5_drug}
\usage{
queryDgidb_V5_drug(drugs)
}
\arguments{
\item{drugs}{character vector, the query drug names.}
}
\description{
a wrapper to query DGIdb API with drug name. Use ghql GraphqlClient function to query DGIdb version 5 to find the interaction between drug and gene. Informations including: "interactionScore","interactionTypes","interactionAttributes","publications", "sources","gene.name","gene.conceptId","gene.longName" ,"Drug".
}
\examples{
use_drugs<-c("Dasatinib","DOVITINIB")
res<-queryDgidb_V5_drug(drugs=use_drugs)

}
