% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{queryDgidb_V5_gene}
\alias{queryDgidb_V5_gene}
\title{queryDgidb_V5_gene}
\usage{
queryDgidb_V5_gene(genes)
}
\arguments{
\item{genes}{character vector, the query gene names.}
}
\description{
a wrapper to query DGIdb API with gene name. Use ghql GraphqlClient function to query DGIdb version 5 to find the interaction between drug and gene, information including: "interactionScore","interactionTypes","interactionAttributes" ,"publications","sources","drug.name","drug.conceptId","queryGene"
}
\examples{
use_genes<-c("ATM","BCL2L1","LCK","CDK4")
res<-queryDgidb_V5_gene (genes=use_genes)

}
