% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{Repurposing_HUB_drugphase_levels}
\alias{Repurposing_HUB_drugphase_levels}
\title{Repurposing_HUB_drugphase_levels}
\usage{
Repurposing_HUB_drugphase_levels()
}
\description{
the drug phases used in Repurposing HUB. Users can select sepecific drug phase
}
\examples{
Repurposing_HUB_drugphase_levels()
}
