% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{ROCTest}
\alias{ROCTest}
\title{ROCTest}
\usage{
ROCTest(eset, G0, G1, verbose = TRUE)
}
\value{
Returns a 'predictive power' (abs(AUC-0.5) * 2) ranked matrix of
putative differentially expressed genes/driver.
}
\description{
An AUC value of 1 means that expression values for this gene alone can perfectly classify the two groupings (i.e. Each of the samples in G1 exhibit a higher level than each of the samples in G2). An AUC value of 0 also means there is perfect classification, but in the other direction. A value of 0.5 implies that the gene has no predictive power to classify the two groups.
}
\details{
ROC-based driver test
}
\examples{
pair.AUC<-list()
single.AUC<-list()
for(i in 1:dim(comps)[1]){
 comp_name <- sprintf('\%s.Vs.\%s',comps[i,1],comps[i,2]) ## each comparison must give a name !!!
 G0  <- rownames(phe_info)[which(phe_info$group==comps[i,2])] # get sample list for G0
 G1  <- rownames(phe_info)[which(phe_info$group==comps[i,1])] # get sample list for G1
 pair.AUC[[comp_name]]<-ROCTest(eset=SINBA.par$SINBA_AC.eset,G0=G0,G1=G1)
 single.AUC[[comp_name]]<-ROCTest(eset=SINBA.par$single_AC.eset,G0=G0,G1=G1)}
\dontrun{
}
}
\concept{DE}
