% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{seed2partners_ftest}
\alias{seed2partners_ftest}
\title{seed2partners_ftest}
\usage{
seed2partners_ftest(
  SINBA_originalID = NULL,
  target_list = NULL,
  N,
  padjust_method = "BH"
)
}
\arguments{
\item{SINBA_originalID}{character vector, the seed-partner driver pairs, seed gene name and partner gene name are seperated by colon. \code{SINBA_originalID} can be manually created or generated by \code{combineNet2target_list} with return_type="SINBA_originalID".}

\item{target_list}{list, the driver-to-target list object. \code{target_list} is generated by \code{get.combined.network} and contains all the driver gene regulons.}

\item{N}{numeric, total unique genes in the whole network or the target_list.}

\item{padjust_method}{character,the type of p value correction methods. User can choose from c ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Default is "BH".}
}
\value{
return a table of fisher exact test.
}
\description{
This function will perform fisher exact test with seed and partner driver regulons.
SINBA_originalID,target_list,N,padjust_method=
}
\examples{
sinba_target_list<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN"),return_type="SINBA_target_list")
sinba_originalID<-combineNet2target_list(target_list=SINBA.par$combined.network$target_list,seed_name="LCK",partner_name=c("BCL2","FYN"),return_type="SINBA_originalID")
sinba_originalID<-as.character(sinba_originalID$originalID)
N<-length(unique(c(SINBA.par$combined.network$network_dat$source,SINBA.par$combined.network$network_dat$target)))

f_df<-seed2partners_ftest(SINBA_originalID=sinba_originalID,target_list=SINBA.par$combined.network$target_list,N=N)
\dontrun{
}
}
