% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim_sinba}
\alias{sim_sinba}
\title{Simulate A Pair of Traits Under the Sinba Model}
\usage{
sim_sinba(tree, Q, births = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{Q}{Transition matrix for simulation.}

\item{births}{A list with the two birth events.
Each element of the list must contain the field
\code{node} with the node at the end of the edge
in which the event happens.
Optionally it can have a field \code{age}
with the time from the start of the edge
in which the event happens.
If not defined,
the age will selected at random point of the edge.
One of the nodes in the birth pair
must be an ancestor of other node.
If it is NULL,
a random pair of birth events will be selected.}
}
\value{
A list with the tree,
births,
and Q matrix used for the simulation,
and factors x and y
with the simulated states on the tips.
}
\description{
\code{sim_sinba} simulates a pair of traits
for a given phylogenetic tree
a given Q transition matrix,
and a pair of birth events.
}
