% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinba-package.R
\docType{package}
\name{sinba-package}
\alias{sinba}
\alias{sinba-package}
\title{sinba: Single Birth of a Trait Model for Correlation Between Discrete Traits in a Phylogeny}
\description{
Package sinba implements the Single Birth of a Trait model to infer the correlation between two discrete traits in a phylogenetic tree.
}
\author{
\strong{Maintainer}: J. Salvador Arias \email{jsalarias@gmail.com} (\href{https://orcid.org/0000-0002-3717-435x}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sergei Tarasov \email{sergei.tarasov@helsinki.fi}
}

}
\keyword{internal}
