% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{SINBA.analysis.dir.create}
\alias{SINBA.analysis.dir.create}
\title{SINBA.analysis.dir.create}
\usage{
SINBA.analysis.dir.create(
  project_main_dir = NULL,
  project_name = NULL,
  tf.network.file = NULL,
  sig.network.file = NULL
)
}
\arguments{
\item{project_main_dir}{character, name or absolute path of the main working directory.}

\item{project_name}{character, name of the project folder.}

\item{tf.network.file}{character, the path of the TF network file.}

\item{sig.network.file}{character, the path of the SIG network file.}
}
\description{
\code{SINBA.analysis.dir.create} is used to help users create an organized working directory for SINBA analysis.However, it is not essential for the analysis.It creates a hierarchcial working directory and returns a list contains this directory information.

This function needs users to define the main working directory and the project's name.
It creates a main working directory with a subdirectory of the project.
It also automatically creates three subfolders (QC, DATA and PLOT) within the project folder.

QC/, storing Quality Control related plots; DATA/, saving data in RData format; PLOT/, storing the plot files generated during the analysis
}
\examples{

\dontrun{
# Creating a main working directory under the current working directory by folder name
SINBA.par <- SINBA.analysis.dir.create("MyMainDir","MyProject")
}
}
