% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{SINBA.loadRData}
\alias{SINBA.loadRData}
\title{SINBA.loadRData}
\usage{
SINBA.loadRData(SINBA.par = NULL, step = "pre-load")
}
\arguments{
\item{SINBA.par}{list, stores all related datasets from driver selection and driver combination prediction pipeline step.}

\item{step}{character, name of the pipeline step decided by user for easier reference.It should be previously assigned by user when calling \code{SINBA.saveRData} function.}
}
\description{
\code{SINBA.loadRData} is a function reloads RData saved by \code{SINBA.saveRData} function. It prevents user from repeating former pipeline steps.
}
\examples{
\dontrun{
SINBA.par <- list()
SINBA.par$out.dir.DATA <- system.file('demo1','/DATA/',package = "SINBA")
SINBA.loadRData(SINBA.par=SINBA.par,step='pre-load')
SINBA.saveRData(SINBA.par=SINBA.par,step='pre-load')
}
}
