% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_fixed_births}
\alias{fit_fixed_births}
\title{Maximum Likelihood Estimation of a Transition Matrix With Fixed Births}
\usage{
fit_fixed_births(tree, data, births, model = NULL, ev_prob = 1, opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{births}{A list with two element,
each element being a list that define a birth event,
with a fields \code{node} indicating the birth of the trait
and \code{age} indicating the time from the start of the edge
in which the event happens.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{ev_prob}{set the probability of a birth event.
By default is 1
(i.e., we have observed different states in the traits).}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_fixed_births()} searches for the maximum likelihood estimate
of the transition matrix
with the Sinba model with two traits
with a fixed birth events.
}
