% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_fixed_births}
\alias{fit_fixed_births}
\title{Maximum Likelihood Estimation of a Transition Matrix With Fixed Births}
\usage{
fit_fixed_births(
  tree,
  data,
  births,
  model = NULL,
  root = NULL,
  root_method = "FitzJohn",
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{births}{A list with two element,
each element being a list that define a birth event,
with a fields \code{node} indicating the birth of the trait
and \code{age} indicating the time from the start of the edge
in which the event happens.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}

\item{root_method}{Method for root calculation at the root.
By default it use the root prior.
If set as "FitzJohn" it will use the FitzJohn et al. (2009)
method,
in which ancestral states are weighted by its own likelihood.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_fixed_births()} searches for the maximum likelihood estimate
of the transition matrix
with the Sinba model with two traits
with a fixed birth events.
}
