% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_fixed_matrix}
\alias{fit_fixed_matrix}
\title{Maximum Likelihood Estimation of Births With Fixed Transition Matrix}
\usage{
fit_fixed_matrix(
  tree,
  data,
  rate_mat,
  model = NULL,
  ev_prob = 1,
  root = "",
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{ev_prob}{Set the probability of a birth event.
By default is 1
(i.e., we have observed different states in the traits).}

\item{root}{Set the root state.
By default,
the root state will be optimized as a parameter.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_fixed_matrix()} searches for the maximum likelihood estimate
of birth events of two traits
given a transition matrix
under the Sinba model.
}
