% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_fixed_matrix}
\alias{fit_fixed_matrix}
\title{Maximum Likelihood Estimation of Births With Fixed Transition Matrix}
\usage{
fit_fixed_matrix(
  tree,
  data,
  rate_mat,
  model = NULL,
  root = NULL,
  root_method = "prior",
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{model}{A model build with \code{new_model()}
or \code{new_hidden_model()}.
By default it uses the independent model.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}

\item{root_method}{Method for root calculation at the root.
By default it use the root prior.
If set as "FitzJohn" it will use the FitzJohn et al. (2009)
method,
in which ancestral states are weighted by its own likelihood.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_fixed_matrix()} searches for the maximum likelihood estimate
of birth events of two traits
given a transition matrix
under the Sinba model.
}
