% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{fit_pagel}
\alias{fit_pagel}
\title{Maximum Likelihood Estimation of the Pagel's Model Under the Sinba Model}
\usage{
fit_pagel(
  tree,
  data,
  model = "IND",
  root = NULL,
  root_method = "prior",
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{Model of evolution for the traits.
By default it uses the independent model ("IND").
The standard model for correlated traits
can be referred as "CORR", "DEP", "xy", or "ARD".
In the "ER" model both traits have equal rates
in any direction;
the "ER2" model also has equal rates,
but rates are different for each trait;
in the "ERs" model the rates of state transitions are equal
but can be different depending on the state.
If the "SYM" model changes between states are equal.
There a two full dependant models,
"x" for a model in which trait x depends on y;
and "y" in which trait y depends on x.
The "coll" model collapse (i.e., removes)
entries for unobserved traits.
In the "sCORR" model,
rates are correlated by the state of the other trait.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}

\item{root_method}{Method for root calculation at the root.
By default it use the root prior.
If set as "FitzJohn" it will use the FitzJohn et al. (2009)
method,
in which ancestral states are weighted by its own likelihood.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_pagel()} searches for the maximum likelihood estimate
of a given model for two traits
using the Pagel (1994) model.
It is a wrapper of \code{fit_fixed_births()}
with the births at the root.
}
