% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel.r
\name{fit_pagel}
\alias{fit_pagel}
\title{Maximum Likelihood Estimation of the Pagel's Model Under the Sinba Model}
\usage{
fit_pagel(tree, data, model = "IND", root = NULL, opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{Model of evolution for the traits.
By default it uses the independent model.
Other valid values are "ARD"
or "xy" for a fully correlated model;
"ER" for a model in which both traits have equal rates
in any direction;
"ER2" for an equal rates model,
but rates are different for each trait;
"SYM" for the symmetric model
in which changes between states are equal;
"x" for a model in which trait x depends on y;
and "y" in which trait y depends on x.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_pagel()} searches for the maximum likelihood estimate
of a given model for two traits
using the Pagel (1994) model.
It is a wrapper of \code{fit_fixed_births()}
with the births at the root.
}
