% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_sinba}
\alias{fit_sinba}
\title{Maximum Likelihood Estimation of the Sinba Model Parameters}
\usage{
fit_sinba(tree, x, y, model = "IND", fixed = NULL, reps = 100, opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{x}{A vector of phenotypic values for a binary trait
for the tips in \code{tree}.}

\item{y}{A vector of phenotypic values for a binary trait
for the tips in \code{tree}.}

\item{model}{Model of evolution for the traits.
By default it uses the independent model.
Other valid values are "ARD"
or "xy" for a fully correlated model;
"ER" for a model in which both traits have equal rates
in any direction;
"ER2" for an equal rates model,
but rates are different for each trait;
"SYM" for the symmetric model
in which changes between states are equal;
"x" for a model in which trait x depends on y;
and "y" in which trait y depends on x.}

\item{fixed}{Use a transition matrix with fixed values.}

\item{reps}{Number of replicates for the Monte Carlo integration.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\value{
An object of class "fit_sinba",
or,
if a \code{fixed} Q matrix is defined,
and object of class "fixed_sinba".
}
\description{
\code{fit_sinba()} searches for the maximum likelihood estimate
of a given model for two traits,
assuming a single birth for each trait.
}
