% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit_sinba}
\alias{fit_sinba}
\title{Maximum Likelihood Estimation of the Sinba Model}
\usage{
fit_sinba(
  tree,
  data,
  model = NULL,
  single_birth = FALSE,
  ev_prob = 1,
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
By default it uses the independent model.}

\item{single_birth}{If true,
it will set the birth of both traits in the same location
(reducing one parameter).}

\item{ev_prob}{set the probability of a birth event.
By default is 1
(i.e., we have observed different states in the traits).}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_sinba()} searches for the maximum likelihood estimate
with the Sinba model for two traits.
It estimate the values of the parameters for the Q matrices
as well as the birth events.
}
