% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.r
\name{fit_sinba_births}
\alias{fit_sinba_births}
\title{Maximum Likelihood Estimation of the Sinba Model Parameters Including Births}
\usage{
fit_sinba_births(
  tree,
  x,
  y,
  model = "IND",
  fixed = NULL,
  births = NULL,
  opts = NULL
)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{x}{A vector of phenotypic values for a binary trait
for the tips in \code{tree}.}

\item{y}{A vector of phenotypic values for a binary trait
for the tips in \code{tree}.}

\item{model}{Model of evolution for the traits.
By default it uses the independent model.
Other valid values are "ARD"
or "xy" for a fully correlated model;
"ER" for a model in which both traits have equal rates
in any direction;
"ER2" for an equal rates model,
but rates are different for each trait;
"SYM" for the symmetric model
in which changes between states are equal;
"x" for a model in which trait x depends on y;
and "y" in which trait y depends on x.}

\item{fixed}{Use a transition matrix with fixed values.}

\item{births}{A list with the two birth events.
Each element of the list must contain the fields
\code{node} with the node at the end of the edge
in which the event happens,
and \code{age} the time from the start of the edge
in which the event happens.
One of the nodes in the birth pair
must be an ancestor of other node.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\value{
An object of class "fit_sinba_births",
or,
if \code{fixed} Q matrix
and \code{birth} list are defined,
an object of class "fixed_sinba_births".
}
\description{
\code{fit_sinba_births()} searches for the maximum likelihood estimate
of a given model for two traits,
assuming a single birth for each trait.
Different from \code{fit_sinba()}
in this function the birth events
are also taken as parameters to be optimized.
}
