% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.r
\name{fit_sinba_single}
\alias{fit_sinba_single}
\title{Maximum Likelihood Estimation of the Sinba Model in a Single Trait}
\usage{
fit_sinba_single(tree, data, root = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
A second should contain the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}
}
\description{
\code{fit_sinba_single()} searches for the maximum likelihood estimate
with the Sinba model
for a single trait.
}
