% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.r
\name{fit_sinba_single}
\alias{fit_sinba_single}
\title{Maximum Likelihood Estimation of the Sinba Model in a Single Trait}
\usage{
fit_sinba_single(tree, data, model = NULL, ev_prob = 1, root = "", opts = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
A second should contain the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{model}{A model build with \code{new_model()},
\code{new_hidden_model()},
or \code{new_rates_model()}.
The model must be defined for a single trait.}

\item{ev_prob}{Det the probability of a birth event.
By default is 1
(i.e., we have observed different states in the traits).}

\item{root}{Set the root state.
By default,
the root state will be optimized as a parameter.}

\item{opts}{User defined parameters for the optimization
with the \code{nloptr} package.
By default it attempts a reasonable set of options.}
}
\description{
\code{fit_sinba_single()} searches for the maximum likelihood estimate
with the Sinba model
for a single trait.
}
