% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fixed_sinba}
\alias{fixed_sinba}
\title{Estimate the Likelihood for a Given Sinba Transition Matrices in Two Traits}
\usage{
fixed_sinba(tree, data, rate_mat, births, semi_mat = NULL, root = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{data}{A data frame with the data.
The first column should contain the taxon names,
The second and third column contains the data,
coded as 0 and 1.
Any other column will be ignored.}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{births}{A list with two element,
each element being a list that define a birth event,
with a fields \code{node} indicating the birth of the trait
and \code{age} indicating the time from the start of the edge
in which the event happens.}

\item{semi_mat}{Rate matrix for the traits with the semi-active process.
If it is NULL,
it will use the same parameter values
from the full process.}

\item{root}{Root prior probabilities.
By default,
all states will have the same probability.}
}
\description{
\code{fixed_sinba()} calculates the likelihood
for a given pair of transition matrices
for two traits
under the Sinba model.
}
