% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{new_rates_model}
\alias{new_rates_model}
\title{Create a New Model With Hidden Rates}
\usage{
new_rates_model(model, rates)
}
\arguments{
\item{model}{A model definition as build with \code{new_model()}.}

\item{rates}{The labels to identify the rates.}
}
\description{
\code{new_rates_model()} creates a new model matrix
with hidden rates.
}
