% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim_sinba}
\alias{sim_sinba}
\title{Simulate A Pair of Traits Under the Sinba Model}
\usage{
sim_sinba(tree, rate_mat, semi_mat = NULL, births = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class "phylo".}

\item{rate_mat}{Rate matrix for the traits with the full process.}

\item{semi_mat}{Rate matrix for the traits with the semi-active process.}

\item{births}{A list with the two birth events.
Each element of the list must contain the field
\code{node} with the node at the end of the edge
in which the event happens.
Optionally it can have a field \code{age}
with the time from the start of the edge
in which the event happens.
If not defined,
the age will selected at random point of the edge.
One of the nodes in the birth pair
must be an ancestor of other node.
If it is NULL,
a random pair of birth events will be selected.}
}
\value{
A list with the tree,
a data frame with the data,
the births,
and the Q and semi-active
matrices used for the simulation.
}
\description{
\code{sim_sinba} simulates a pair of traits
for a given phylogenetic tree
a given Q transition matrix,
and a pair of birth events.
}
